split team runtime tool wrappers
